pyproject.toml
[tool.poetry]
name = "Read-PDF-API"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.3.1"}
pydantic = ">=2.7"
langchain = "^0.3.20"
langchain-community = "^0.3.19"
pdfminer-six = "20231228"
langchain-ollama = "^0.3.2"
faiss-cpu = "^1.10.0"
pdfplumber = "0.11.3"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
